Add garbage collector for hvml
The chosen implementation is a depth-first generational garbage collector.

I've experimented with a couple of designs, but this was what performed best
while still being simple (~60 LOC).

It uses a depth-first graph traversal strategy. The heap is separated into 2
sections, new generation and old generation. Terms are always allocated in the
new generation. When the GC runs a minor collection, it moves terms from the new
area to the old area, compressing the memory in the process.

When the old generation grows too much, we also run a major collection. It
collects all the memory and not just the new roots by swapping the new and old
regions. The entire memory is then treated as new and collected. This process then
repeats itself, although for the 16GB M4 mac mini it won't run much more than 2 major
collections before starting to thrash the swap.

Three new structures were added to the State:
- Relocation table, which we use to store the Terms that are outside global memory,
like the local Haskell variables that need to be updated with the GC. Since they
are used lazily I made the table use a use count, with terms becoming "inactive"
after the counter gets to 0.
- Old generation update buffer. When doing minor GC, old gen terms that pointed to
new gen ones need to be updated. Whenever we `set` and old gen location to point to a
new gen one we add that location to this queue. It gets emptied after each minor GC.
- GC stack. Used to implement depth-first traversal of the graph. This is the fastest
implementation of graph traversal from what I tested, even faster than O(1) breadth-first,
probably due to better cache use.

The GC parameters were chosen to perform well for SupGen programs.
I chose to run minor GCs every 2^29 mem allocation or 4GB. The first run is delayed
to 2^30 or 8GB to avoid the big performance hit of the GC for smaller programs.
The major GC runs when the old gen memory passes 2^29 elements as well.

In general, not accounting for memory pressure and swapping, everything runs
around 10-20% slower due to needing to keep track of some GC information right in the
middle of the hottest of paths (the get function).
Since the first GC is so big, between the first and the second, performance is ~30-50% worse.
After the second GC we start to gain performance back due to better memory use and
eventually because we don't need to swap so soon.
The first major GC also gives a big performance loss while not recuperating that much
memory, only ~30% compared to ~70% for minor GCs. If we don't want to bother with much
longer programs (>20 execution time for current supgen) it might be better to just
remove major collections.