Set the terms in the stack when returning from reduce after reaching a VAR
When we finished a C reduce call because we reached an unsubstituted variable,
the reduce function would set the last term in the stack to point to that variable and return.

In some very rare cases, it could happen that just setting the last value of the stack
is not enough. The correct behaviour is to instead set all the values values in the stack
to the next value in the stack, guaranteeing that every Term points to it's correct successor.
If only the last value is set it could happen that the other values in the stack before it
point to the wrong locations (or rather that the locations they point to contain

This bug started occuring after the 1-port LAM optimization, when we stopped writing the full
result of the LAM substitution to the HEAP and instead only wrote the substituted variable value,
and it got worse (occurs in more cases) when the equivalent optimization was done to DUPs.

The simplest example I found is the following church exponentiation 2^3:
```
@main = (
  λf !&1{f1 f}=f !&1{f2 f}=f λx (f1 (f2 (f x)))
  λf !&0{f1 f}=f λx (f1 (f x)))
```

This is a rather long example, with 27 interactions to normalization, but at step 15 we have the
following subnet:
```
!&0{a b} = c
!&1{c d} = &1{(... ...) (... ...)}
```
With the stack at this moment being:
```
term_new(SUB,0x0000,0x000000001a)
term_new(DP0,0x0000,0x000000002f)
term_new(DP0,0x0001,0x000000002a)
```
After this interaction, we write the second app to address 0x2a and return
the first app to the stack in the reduce function.
The reduction continues, at step 18 we annihilate the APP returned to the stack in step 15 and
then reach an unsubstituted variable indicating the end of this call to the reduce function.

At that moment the stack contains:
```
term_new(SUB,0x0000,0x000000001a)
term_new(DP0,0x0000,0x000000002f)
term_new(DP0,0x0000,0x000000003b)
term_new(APP,0x0000,0x0000000039)
term_new(DP0,0x0000,0x0000000029)
```

`reduce` returns the DP0 pointing to 0x2f and the extract function in haskell will continue
to normalize the value of the Dup. However, at this point the memory location 0x2f does not
contain the current value of the Dup, which should the DP0 pointing to 0x3b, instead it contains
that second APP that was written to the heap of the Dup in step 15.

Here we can see why it is necessary to update the values that are stored in the stack
when returning from the reduce function.