'&var' now means a variable is auto duplicated
This commit gives a new meaning to variables declared with '&' in front of their names.
Previously, we used that for dup/sup labels received as function arguments.

Now, any variable that is declared with '&' in front of its name will be auto-duplicated with unique labels.
The '&' goes only in the bind location and not in any of its uses.
The repeated names are transformed from 'name' into 'name_dupN' and duplications are added at the top
of the body of the binding term.

Example:
The fn `@foo = λ&x λp (p x x x)` will become `@foo = λx !&{x x_dup1} = x; !&{x_dup1 x_dup2} = x_dup1; (p x x_dup1 x_dup2)`.

Now any variable can be used as superposition label, even those without '&' in front of their names.
'&' arguments are no longer strict, the strictness annotation is needed in front of the dup annotation.

For example, if our function used to be `@foo(&l x) = ...`, the new equivalent term is `@foo(!&l x) = ...`,
and the `&` is no longer part of the name, the variable is simply `l`.

For matches, the move clause needs to specify if the moved variable is linear or non-linear as well.
For example, in this term, `b` is linear before the match and non-linear (as separate variables) in each match case.
`@foo(a b) = ~ a !&b { 0: 1; 1+p: (+ p (+ b b)) }`

It can also be the other way around, if a variable is bound as non-linear but not marked as non-linear in the move, it'll be linear inside the match.
For example, this will be an error: `@foo(a &b) = ~ a !b { 0: b; 1+a: (+ b b) }`.